backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
tavily-python
google-api-python-client
openai
apscheduler
pyyaml
python-dotenv
pytest
httpx
